services/esp32_sim/requirements.txt
paho-mqtt==2.0.0
python-dateutil==2.8.2
pandas==2.1.4